+++ src/Editor.js
@@ -10,7 +10,9 @@ import * as UndoRedoManager from './model/UndoRedoManager';
 import * as ModelStats from './util/ModelStats';
 import * as ImageRenderer from './renderer/canvas/ImageRenderer';
 import * as RecognizerContext from './model/RecognizerContext';
+import * as SmartGuide from './smartguide/SmartGuide';
 import Constants from './configuration/Constants';
+
 
 /**
  * Trigger callbacks
@@ -131,6 +133,11 @@ function manageRecognizedModel(editor, model, ...types) {
     triggerCallbacks(editor, undefined, ...types);
   }
 
+  if (editor.configuration.recognitionParams.type === 'TEXT' && editor.configuration.recognitionParams.apiVersion === 'V4' && editor.configuration.recognitionParams.v4.text.smartGuide) {
+    // eslint-disable-next-line no-use-before-define
+    launchSmartGuide(editorRef, modelRef.exports);
+  }
+
   if ((InkModel.extractPendingStrokes(model).length > 0) &&
     (!editor.recognizer.addStrokes) && // FIXME: Ugly hack to avoid double export (addStrokes + export)
     (editor.configuration.triggers.exportContent !== Constants.Trigger.DEMAND)) {
@@ -200,6 +207,16 @@ function addStrokes(editor, model, trigger = editor.configuration.triggers.addSt
         }
       });
   }
+}
+
+/**
+ * Launch smartguide.
+ * @param {Editor} editor
+ * @param {Object} exports
+ */
+function launchSmartGuide(editor, exports) {
+  const editorRef = editor;
+  editorRef.smartGuide = SmartGuide.launchSmartGuide(editor.smartGuide, exports);
 }
 
 /**
@@ -295,6 +312,7 @@ function launchResize(editor, model) {
           });
         }, editor.configuration.resizeTriggerDelay);
       });
+    SmartGuide.insertSmartGuide(editor.smartGuide);
   }
 }
 
@@ -396,6 +414,8 @@ export class Editor {
 
     this.theme = theme;
     this.penStyle = penStyle;
+
+    this.smartGuide = SmartGuide.createSmartGuide(this);
 
     /**
      * @private