+++ src/configuration/Constants.js
@@ -53,7 +53,8 @@ const Constants = {
     RENDERER: 'renderer',
     RECOGNIZER: 'recognizer',
     CALLBACK: 'callback',
-    UTIL: 'util'
+    UTIL: 'util',
+    SMARTGUIDE: 'smartguide'
   },
   LogLevel: {
     TRACE: 'TRACE',
@@ -65,6 +66,9 @@ const Constants = {
   Error: {
     NOT_REACHABLE: 'MyScript recognition server is not reachable. Please reload once you are connected.',
     WRONG_CREDENTIALS: 'Application credentials are invalid. Please check or regenerate your application key and hmackey.'
+  },
+  Exports: {
+    JIIX: 'application/vnd.myscript.jiix'
   }
 };
 export default Constants;